+++ src/lib.rs
@@ -27,8 +27,10 @@ pub mod samples;
 #[cfg(feature = "simulator")]
 pub mod simulator;
 pub mod songs;
+pub mod state;
 #[cfg(test)]
 pub mod testutil;
 pub mod trigger;
+pub mod tui;
 pub mod util;
 pub mod verify;